ユーザーの登録、編集、削除機能を実装

+++ src/main/java/kensyu3/controller/QuestionsAnswersAction.java
@@ -28,7 +28,8 @@ public class QuestionsAnswersAction extends Base{
 	private int[] questionsId;
 	private String inputQuestion;
 	private String[] inputAnswers;
-	private UsersBean user = new UsersBean();
+	private String currentUserName;
+	private int currentUserAuth = 2;
 	private String errorMessage;
 	private int point;
 	private int correctCnt;
@@ -37,13 +38,24 @@ public class QuestionsAnswersAction extends Base{
 	/*
 	 *  アクションの定義
 	 */
+	//top画面を表示
+	public String top() throws Exception{
+		//Baseクラスでログインしているかどうかを確認
+		if (super.isCheckLogin()) {
+			//top画面に遷移
+			return "top" ;
+		}else {
+			//login画面に遷移
+			return "login";
+		}
+	}
 	
 	//list画面を表示
 	public String list() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
 			//権限があるかどうか確認
-			if(getUser().getAdminFlag() == 1) {
+			if(getCurrentUserAuth() == 1) {
 				//登録されている問題がある場合
 				if(!(getQueList().isEmpty())) {
 					//list画面に遷移
@@ -58,28 +70,34 @@ public class QuestionsAnswersAction extends Base{
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
-	//登録画面を表示
+	//register画面を表示
 	public String register() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			//register画面に遷移
-			return "success" ;
+			//権限があるかどうか確認
+			if(getCurrentUserAuth() == 1) {
+				//register画面に遷移
+				return "register" ;
+			}else {
+				//error画面に遷移
+				return "error";
+			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
-	//登録確認画面を表示
+	//register_confirm画面を表示
 	public String register_confirm() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			//登録画面を経由したか
-			if(inputQuestion != null && inputAnswers != null) {
+			//権限と登録画面を経由したかを確認
+			if(getCurrentUserAuth() == 1 && inputQuestion != null) {
 				Validation val = new Validation();
 				//問題や答えにエラーがないか確認し、エラーメッセージに値を格納
 				errorMessage = val.validate(inputQuestion, inputAnswers);
@@ -91,7 +109,7 @@ public class QuestionsAnswersAction extends Base{
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
@@ -99,26 +117,32 @@ public class QuestionsAnswersAction extends Base{
 	public String register_complete() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			QuestionsDao queDao = new QuestionsDao();
-			AnswersDao ansDao = new AnswersDao(); 
-			//問題を登録
-			int questionId = queDao.register(inputQuestion);
-			//答えを登録
-			ansDao.register(questionId, inputAnswers);
-			//list画面に遷移
-			return "success";
+			//権限と登録確認画面を経由したか確認
+			if(getCurrentUserAuth() == 1 && inputQuestion != null) {
+				QuestionsDao queDao = new QuestionsDao();
+				AnswersDao ansDao = new AnswersDao(); 
+				//問題を登録
+				int questionId = queDao.register(inputQuestion);
+				//答えを登録
+				ansDao.register(questionId, inputAnswers);
+				//list画面に遷移
+				return "list";
+			}else {
+				//error画面に遷移
+				return "error";
+			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
-	//削除確認画面を表示
+	//delete_confirm画面を表示
 	public String delete_confirm() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			//パラメータで指定したquestionIdが存在するか
-			if(getQuestion() != null) {
+			//権限とパラメータで指定したquestionIdが存在するか確認
+			if(getCurrentUserAuth() == 1 && getQuestion() != null) {
 				//delete_confirm画面に遷移
 				return "delete_confirm";
 			}else {
@@ -127,7 +151,7 @@ public class QuestionsAnswersAction extends Base{
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
@@ -135,31 +159,36 @@ public class QuestionsAnswersAction extends Base{
 	public String delete_complete() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			QuestionsDao queDao = new QuestionsDao();
-			AnswersDao ansDao = new AnswersDao();
-			//問題と答えを削除
-			queDao.delete(questionId);
-			ansDao.deleteByQuestionId(questionId);
-			//登録されている問題がある場合
-			if(!(getQueList().isEmpty())) {
-				//list画面に遷移
-				return "list" ;
+			//権限とdelete_confirm画面を経由したか確認
+			if(getCurrentUserAuth() == 1 && questionId != 0) {
+				QuestionsDao queDao = new QuestionsDao();
+				AnswersDao ansDao = new AnswersDao();
+				//問題と答えを削除
+				queDao.delete(questionId);
+				ansDao.deleteByQuestionId(questionId);
+				//削除後、登録されている問題がある場合
+				if(!(getQueList().isEmpty())) {
+					//list画面に遷移
+					return "list" ;
+				}else {
+					//top画面に遷移
+					return "top";
+				}
 			}else {
-				//top画面に遷移
-				return "top";
+				return "error";
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
-	//編集画面を表示
+	//edit画面を表示
 	public String edit() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			//パラメータで指定したquestionIdが存在するか
-			if(getQuestion() != null) {
+			//権限とパラメータで指定したquestionIdが存在するか確認
+			if(getCurrentUserAuth() == 1 && getQuestion() != null) {
 				//edit画面に遷移
 				return "edit";
 			}else {
@@ -168,16 +197,16 @@ public class QuestionsAnswersAction extends Base{
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
-	//編集確認画面を表示
-	public String edit_confirm(){
+	//edit_confirm画面を表示
+	public String edit_confirm() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			//編集画面を経由したか
-			if(inputQuestion != null && inputAnswers != null) {
+			//権限と編集画面を経由したか確認
+			if(getCurrentUserAuth() == 1 && inputQuestion != null) {
 				Validation val = new Validation();
 				//問題や答えにエラーがないか確認し、エラーメッセージに値を格納
 				errorMessage = val.validate(inputQuestion, inputAnswers);
@@ -189,7 +218,7 @@ public class QuestionsAnswersAction extends Base{
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
@@ -197,39 +226,45 @@ public class QuestionsAnswersAction extends Base{
 	public String edit_complete() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			QuestionsDao queDao = new QuestionsDao();
-			AnswersDao ansDao = new AnswersDao();
-			//問題を編集
-			queDao.update(inputQuestion,questionId);
-			
-			//問題idからDBに登録されている答えデータを取得
-			ArrayList<AnswersBean> aList = ansDao.findByQuestionId(questionId);
-			
-			//入力された答えの数だけ処理を繰り返す
-			for(int i = 0; i < inputAnswers.length; i++) {
-				if( i < answersId.length) { //入力された答えの中に、idを持つものがあった場合（更新された答えがあった場合）
-					//idをもとにDBに登録されている答えデータを取得
-					AnswersBean tmpAnswer = ansDao.findById(answersId[i]);
-					//入力された答えと、DBに登録されている答えの内容が一致していなかった場合
-					if(!(inputAnswers[i].equals(tmpAnswer.getAnswer()))) {
-						ansDao.update(answersId[i], inputAnswers[i]); //答えを更新
-					}
-				} else { //idを持たない答えがあった場合（新たに追加された答えがあった場合）
-					ansDao.register(questionId, inputAnswers[i]); //答えを登録
-				}
-			}
-			if(aList.size() > answersId.length) { //既存の答えの数の方が、フォームから渡されたidの数より多かった場合（削除された答えがあった場合）
-				for(AnswersBean ans : aList) { //既存の答えの数だけ、処理を繰り返す
-					if(!(Arrays.stream(answersId).anyMatch(x -> x == ans.getId()))){ //既存の答えにしかないidがあった場合
-						ansDao.deleteById(ans.getId()); //答えを削除
+			//権限と編集確認画面を経由したか確認
+			if(getCurrentUserAuth() == 1 && inputQuestion != null) {
+				QuestionsDao queDao = new QuestionsDao();
+				AnswersDao ansDao = new AnswersDao();
+				//問題を編集
+				queDao.update(inputQuestion,questionId);
+				
+				//問題idからDBに登録されている答えデータを取得
+				ArrayList<AnswersBean> aList = ansDao.findByQuestionId(questionId);
+				
+				//入力された答えの数だけ処理を繰り返す
+				for(int i = 0; i < inputAnswers.length; i++) {
+					if( i < answersId.length) { //入力された答えの中に、idを持つものがあった場合（更新された答えがあった場合）
+						//idをもとにDBに登録されている答えデータを取得
+						AnswersBean tmpAnswer = ansDao.findById(answersId[i]);
+						//入力された答えと、DBに登録されている答えの内容が一致していなかった場合
+						if(!(inputAnswers[i].equals(tmpAnswer.getAnswer()))) {
+							ansDao.update(answersId[i], inputAnswers[i]); //答えを更新
+						}
+					} else { //idを持たない答えがあった場合（新たに追加された答えがあった場合）
+						ansDao.register(questionId, inputAnswers[i]); //答えを登録
 					}
 				}
+				if(aList.size() > answersId.length) { //既存の答えの数の方が、フォームから渡されたidの数より多かった場合（削除された答えがあった場合）
+					for(AnswersBean ans : aList) { //既存の答えの数だけ、処理を繰り返す
+						if(!(Arrays.stream(answersId).anyMatch(x -> x == ans.getId()))){ //既存の答えにしかないidがあった場合
+							ansDao.deleteById(ans.getId()); //答えを削除
+						}
+					}
+				}
+				//list画面に遷移
+				return "list";
+			}else {
+				//error画面に遷移
+				return "error";
 			}
-			//list画面に遷移
-			return "success";
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
@@ -238,10 +273,10 @@ public class QuestionsAnswersAction extends Base{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
 			//test画面に遷移
-			return "success";
+			return "test";
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
@@ -249,7 +284,7 @@ public class QuestionsAnswersAction extends Base{
 	public String result() throws Exception{
 		//Baseクラスでログインしているかどうかを確認
 		if (super.isCheckLogin()) {
-			//テスト画面を経由したか
+			//テスト画面を経由したか確認
 			if(inputAnswers != null) {
 				//入力された答えの数だけ処理を繰り返す
 				for(int i = 0; i < inputAnswers.length; i++) {
@@ -281,7 +316,7 @@ public class QuestionsAnswersAction extends Base{
 			}
 		}else {
 			//login画面に遷移
-			return "failure";
+			return "login";
 		}
 	}
 	
@@ -433,15 +468,29 @@ public class QuestionsAnswersAction extends Base{
 		return correctCnt;
 	}
 	
-	//現在ログインしているユーザーのgetter
-	public UsersBean getUser() throws Exception{
+	//現在ログインしているユーザ名のgetter
+	public String getCurrentUserName() throws Exception{
 		//userNameに値が格納されていない場合
-		if(user.getId() == 0) {
+		if(currentUserName == null) {
 			UsersDao usersDao = new UsersDao();
 			//セッションに登録されているidをもとに、ユーザーを取得
-			user = usersDao.findById((int)session.get("userId"));
+			UsersBean currentUser = usersDao.findById((int)session.get("userId"));
+			currentUserName = currentUser.getName();
 		}
-		return user;
+		return currentUserName;
+	}
+	
+	//現在ログインしているユーザ権限のgetter
+	public int getCurrentUserAuth() throws Exception{
+		//currentUserAuthに0か1が格納されていない場合
+		if(currentUserAuth == 2) {
+			UsersDao usersDao = new UsersDao();
+			//セッションに登録されているidをもとに、ユーザーを取得
+			UsersBean currentUser = usersDao.findById((int)session.get("userId"));
+			//ユーザー情報からidを取得
+			currentUserAuth = currentUser.getAdminFlag();
+		}
+		return currentUserAuth;
 	}
 	
 	//現在日時のgetter
